Factor out permutation utility for testing
